Add --dry-run option

+++ src/main.rs
@@ -32,6 +32,12 @@ fn replace(dest_ptn: &str, substrings: &[String]) -> String {
 fn main() {
     let matches = app_from_crate!()
         .arg(
+            Arg::with_name("dry-run")
+                .short("n")
+                .long("dry-run")
+                .help("Do not actually move the files, just show what would be done."),
+        )
+        .arg(
             Arg::with_name("SOURCE")
                 .required(true)
                 .index(1)
@@ -46,6 +52,7 @@ fn main() {
         .get_matches_from(env::args_os());
     let src_ptn = matches.value_of("SOURCE").unwrap();
     let dest_ptn = matches.value_of("DEST").unwrap();
+    let dry_run = 0 < matches.occurrences_of("dry-run");
 
     let matches = match walk(Path::new("."), src_ptn) {
         Err(err) => {
@@ -78,8 +85,10 @@ fn main() {
         line.push_str(" --> "); //TODO: Wrap line if it's too long
         line.push_str(dest);
         println!("{}", line);
-        if let Err(err) = std::fs::rename(src, dest) {
-            eprintln!("Failed to copy \"{}\": {}", src, err);
+        if !dry_run {
+            if let Err(err) = std::fs::rename(src, dest) {
+                eprintln!("Failed to copy \"{}\": {}", src, err);
+            }
         }
     }
 }
